services/auth_service/requirements.txt
Django
djangorestframework
django-cors-headers
django-environ
gunicorn
psycopg2-binary
djangorestframework-simplejwt
requests